requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
gunicorn==21.2.0
pytz==2024.1

